requirements/dev.txt
# =============================================================================
# Dépendances de développement local LYNEERP.
# =============================================================================

-r base.txt

# Tests
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
Faker==37.12.0

# Lint / qualité
ruff==0.6.9
mypy==1.13.0
django-stubs==5.1.1
djangorestframework-stubs==3.15.1

# Outils dev
django-debug-toolbar==4.4.6
ipython==8.27.0
